+++ themes/theme-rui/src/components/Card.styles.ts
@@ -1,7 +1,8 @@
 import { cva } from '@marigold/system';
 import type { ThemeComponent } from '@marigold/system';
 
-export const Card: ThemeComponent<'Card'> = cva(['p-4 rounded-md'], {
+export const Card: ThemeComponent<'Card'> = cva({
+  base: ['p-4 rounded-md'],
   variants: {
     variant: {
       default: 'ui-surface shadow-elevation-raised',