+++ themes/theme-rui/src/components/SectionMessage.styles.ts
@@ -1,30 +1,29 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const SectionMessage: ThemeComponent<'SectionMessage'> = {
-  container: cva(
-    [
+  container: cva({
+    base: [
       "grid-cols-[min-content_auto_min-content] gap-x-4 gap-y-1 [grid-template-areas:'icon_title_close''icon_content_content']",
       'bg-background rounded-md border px-3 py-4',
     ],
-    {
-      variants: {
-        variant: {
-          success:
-            'border-success-muted-accent bg-success-muted text-success-muted-foreground',
-          warning:
-            'border-warning-muted-accent bg-warning-muted text-warning-muted-foreground',
-          info: 'border-info-muted-accent bg-info-muted text-info-muted-foreground',
-          error:
-            'border-destructive-muted-accent bg-destructive-muted text-destructive-muted-foreground',
-        },
+    variants: {
+      variant: {
+        success:
+          'border-success-muted-accent bg-success-muted text-success-muted-foreground',
+        warning:
+          'border-warning-muted-accent bg-warning-muted text-warning-muted-foreground',
+        info: 'border-info-muted-accent bg-info-muted text-info-muted-foreground',
+        error:
+          'border-destructive-muted-accent bg-destructive-muted text-destructive-muted-foreground',
       },
-      defaultVariants: {
-        variant: 'info',
-      },
-    }
-  ),
-  title: cva('text-sm font-medium'),
-  content: cva('text-muted-foreground text-sm leading-5 font-normal', {
+    },
+    defaultVariants: {
+      variant: 'info',
+    },
+  }),
+  title: cva({ base: 'text-sm font-medium' }),
+  content: cva({
+    base: 'text-muted-foreground text-sm leading-5 font-normal',
     variants: {
       variant: {
         success: 'text-success-muted-foreground',
@@ -37,7 +36,8 @@ export const SectionMessage: ThemeComponent<'SectionMessage'> = {
       variant: 'info',
     },
   }),
-  icon: cva('h-6 w-6 align-baseline leading-none -mt-0.5', {
+  icon: cva({
+    base: 'h-6 w-6 align-baseline leading-none -mt-0.5',
     variants: {
       variant: {
         success: 'text-success-muted-accent',
@@ -50,9 +50,11 @@ export const SectionMessage: ThemeComponent<'SectionMessage'> = {
       variant: 'info',
     },
   }),
-  close: cva([
-    'size-8 text-foreground',
-    '[&_svg]:size-6',
-    '-my-1.5 -me-2', // align button with title
-  ]),
+  close: cva({
+    base: [
+      'size-8 text-foreground',
+      '[&_svg]:size-6',
+      '-my-1.5 -me-2', // align button with title
+    ],
+  }),
 };