+++ themes/theme-rui/src/components/Switch.styles.ts
@@ -1,20 +1,24 @@
 import { type ThemeComponent, cva } from '@marigold/system';
 
 export const Switch: ThemeComponent<'Switch'> = {
-  container: cva(
-    'disabled:cursor-not-allowed disabled:text-disabled-foreground'
-  ),
-  track: cva([
-    'flex h-6 w-10 shrink-0 cursor-pointer items-center rounded-full transition-colors',
-    'border-2 border-transparent', // this increases the hit area so it is 24px
-    'group-disabled/switch:bg-disabled group-disabled/switch:text-disabled-foreground group-selected/switch:group-disabled/switch:bg-disabled group-selected/switch:group-disabled/switch:text-disabled-foreground',
-    'group-selected/switch:bg-brand bg-input',
-    'group-focus-visible/switch:ui-state-focus outline-none',
-  ]),
-  thumb: cva([
-    'pointer-events-none block size-5 rounded-full',
-    'bg-background shadow-xs',
-    'ring-0 transition-transform duration-150 ease-out-quint',
-    'group-selected/switch:translate-x-4 translate-x-0',
-  ]),
+  container: cva({
+    base: 'disabled:cursor-not-allowed disabled:text-disabled-foreground',
+  }),
+  track: cva({
+    base: [
+      'flex h-6 w-10 shrink-0 cursor-pointer items-center rounded-full transition-colors',
+      'border-2 border-transparent', // this increases the hit area so it is 24px
+      'group-disabled/switch:bg-disabled group-disabled/switch:text-disabled-foreground group-selected/switch:group-disabled/switch:bg-disabled group-selected/switch:group-disabled/switch:text-disabled-foreground',
+      'group-selected/switch:bg-brand bg-input',
+      'group-focus-visible/switch:ui-state-focus outline-none',
+    ],
+  }),
+  thumb: cva({
+    base: [
+      'pointer-events-none block size-5 rounded-full',
+      'bg-background shadow-xs',
+      'ring-0 transition-transform duration-150 ease-out-quint',
+      'group-selected/switch:translate-x-4 translate-x-0',
+    ],
+  }),
 };