+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @marigold/icons
 
+## 1.3.33
+
+### Patch Changes
+
+- Updated dependencies [fd1b092]
+  - @marigold/system@17.1.0
+
 ## 1.3.32
 
 ### Patch Changes